requirements.txt
asgiref==3.8.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django>=3.2.4,<3.3
django-allauth==0.63.3
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
whitenoise==6.7.0
